// frontend/utils/types.ts
/////////////////////
// IMPORTS SECTION //
/////////////////////


////////////////////////
// INTERFACES SECTION //
////////////////////////
export type Booking = {
  id: string
  customerName: string
  customerEmail: string
  customerPhone: string
  startDate: string
  endDate: string
  state: string
}